requirements.txt
# Telegram Bot Framework
aiogram==3.13.1

# HTTP Client
httpx==0.27.2

# Environment Variables
python-dotenv==1.0.1

# Google Calendar API
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Timezone Support
pytz==2024.2

# Speech-to-Text (Whisper для локального распознавания речи)
openai-whisper>=20231117
# Требует ffmpeg: Windows - https://ffmpeg.org/download.html, Linux - apt-get install ffmpeg, macOS - brew install ffmpeg
# Требует ~1.5GB места для модели base (скачивается автоматически при первом запуске)

